+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `FlowbiteSlider` and `FlowbiteRangeSlider` components.
 
+### Changed
+- Rename `FlowbiteTheme.extension` static method to `FlowbiteTheme.of`.
+
 ## [0.0.6] - 12-09-2026
 ### Added
 - Add `FlowbiteBreadcrumb` component.