+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.7] - 15-09-2026
 ### Added
 - Add `FlowbiteSlider` and `FlowbiteRangeSlider` components.
 - Add `FlowbiteKbd` component.
@@ -67,7 +67,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `FlowbiteInputField` component.
 - Add `FlowbiteInputLabel` component.
 
-[Unreleased]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.6...dev
+[Unreleased]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.7...dev
+[0.0.7]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/hanmajid/flowbite_ui/compare/v0.0.3...v0.0.4